Finish Project Api

+++ Order.Data/Models/User.cs
@@ -12,8 +12,8 @@ namespace Orders.Data.Models
     {
         [Required]
         public string FullName { get; set; }
-        public decimal? Latitude { get; set; }
-        public decimal? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
         public string Address { get; set; }
         public DateTime CreatedAt { get; set; }
         public bool IsDelete { get; set; }
